+++ samples/eltwise/eltwise_unary_dropout.c
@@ -267,11 +267,11 @@ int test_dropout_fwd( const libxsmm_blasint bitm, const libxsmm_blasint M, const
   libxsmm_blasint mask_ld = (bitm == 0) ? ldo : ((ldo+15)-((ldo+15)%16))/8;
 
   if ( M > ldi ) {
-    fprintf( stderr, "test_dropout_fwd %i %i %i: ldi needs to be equal to or bigger than M\n", dtype_in, dtype_out, dtype_comp);
+    fprintf( stderr, "test_dropout_fwd %i %i %i: ldi needs to be equal to or bigger than M\n", (int)dtype_in, (int)dtype_out, (int)dtype_comp);
     exit(-1);
   }
   if (M > ldo ) {
-    fprintf( stderr, "test_dropout_fwd %i %i %i: ldo needs to be equal to or bigger than N\n", dtype_in, dtype_out, dtype_comp);
+    fprintf( stderr, "test_dropout_fwd %i %i %i: ldo needs to be equal to or bigger than N\n", (int)dtype_in, (int)dtype_out, (int)dtype_comp);
     exit(-1);
   }
 
@@ -335,12 +335,12 @@ int test_dropout_fwd( const libxsmm_blasint bitm, const libxsmm_blasint M, const
     for ( i = 0; i < N; ++i ) {
       for ( j = 0; j < M/8; ++j ) {
         if ( mask_gold[(i*mask_ld)+j] != mask[(i*mask_ld)+j] ) {
-          printf("error at possition i=%i, j=%i, %u, %u\n", i, j, mask[(i*mask_ld)+j], mask_gold[(i*mask_ld)+j]);
+          printf("error at possition i=%i, j=%i, %i, %i\n", i, j, mask[(i*mask_ld)+j], mask_gold[(i*mask_ld)+j]);
           s = 1;
         }
 #if 0
         else {
-          printf("correct at possition i=%i, j=%i, %u, %u\n", i, j, mask[(i*mask_ld)+j], mask_gold[(i*mask_ld)+j]);
+          printf("correct at possition i=%i, j=%i, %i, %i\n", i, j, mask[(i*mask_ld)+j], mask_gold[(i*mask_ld)+j]);
         }
 #endif
       }
@@ -363,9 +363,9 @@ int test_dropout_fwd( const libxsmm_blasint bitm, const libxsmm_blasint M, const
   libxsmm_free( mask_gold );
 
   if ( ret == EXIT_SUCCESS ) {
-    printf("SUCCESS unary dropout fwd %i %i %i\n", dtype_in, dtype_out, dtype_comp);
+    printf("SUCCESS unary dropout fwd %i %i %i\n", (int)dtype_in, (int)dtype_out, (int)dtype_comp);
   } else {
-    printf("FAILURE unary dropout fwd %i %i %i\n", dtype_in, dtype_out, dtype_comp);
+    printf("FAILURE unary dropout fwd %i %i %i\n", (int)dtype_in, (int)dtype_out, (int)dtype_comp);
   }
 
   return ret;
@@ -389,11 +389,11 @@ int test_dropout_bwd( const libxsmm_blasint M, const libxsmm_blasint N, const li
   libxsmm_blasint mask_ld = ((ldi+15)-((ldi+15)%16))/8;
 
   if ( M > ldi ) {
-    fprintf( stderr, "test_dropout_fwd %i %i %i: ldi needs to be equal to or bigger than M\n", dtype_in, dtype_out, dtype_comp);
+    fprintf( stderr, "test_dropout_fwd %i %i %i: ldi needs to be equal to or bigger than M\n", (int)dtype_in, (int)dtype_out, (int)dtype_comp);
     exit(-1);
   }
   if (M > ldo ) {
-    fprintf( stderr, "test_dropout_fwd %i %i %i: ldo needs to be equal to or bigger than N\n", dtype_in, dtype_out, dtype_comp);
+    fprintf( stderr, "test_dropout_fwd %i %i %i: ldo needs to be equal to or bigger than N\n", (int)dtype_in, (int)dtype_out, (int)dtype_comp);
     exit(-1);
   }
